Initializations

Beyond the standard initializations available in Flux there are a couple of new initializations available:

FourierFilterFlux.positive_glorot_uniform — Function

positive_glorot_uniform(dims...)

same idea as a glorot_uniform, but limited to strictly positive entries.

FourierFilterFlux.iden_perturbed_gaussian — Function

iden_perturbed_gaussian(dims...)

an identity along the diagonal with Gaussian deviations of standard deviation $\frac 1 {100}$ everywhere

FourierFilterFlux.uniform_perturbed_gaussian — Function

uniform_perturbed_gaussian(dims...)

If there are $n$ total entries in the matrix, each entry is gaussian distributed with a mean of $¹/ₙ$ and a standard deviation of $\frac{1}{10·n}$