Built-in constructors

This section details the filter constructors included by default. These are built off of the Wavelets and Shearlab packages for the 1D and 2D constructors respectively.

1D methods

FourierFilterFlux.waveletLayer — Function

waveletLayer(inputSize::Union{Int,NTuple{N, T}}; 
             dType = Float32, σ = identity, trainable = false,
             plan = true, init = Flux.glorot_normal, bias=false,
             convBoundary=Sym(), cw = Morlet(), averagingLayer = false,
             varargs...) where {N,T}

Create a ConvFFT layer that uses wavelets from ContinuousWavelets.jl. By default it isn't trainable. varargs are any of the settings that can be passed on to creating a CFW type.

New Arguments

convBoundary::ConvBoundary=Sym(): the type of symmetry to use in computing the transform. Note that convBoundary and boundary are different, with boundary needing to be set using types from ContinuousWavelets and convBoundary needs to be set using the FourierFilterFlux boundary types.

cw::ContWaveClass=Morlet(): the type of wavelet to use, e.g. dog2, Morlet(). See ContinuousWavelets.jl for more.

averagingLayer::Bool=false: the same idea as for shearingLayer, if this is true, only return the results from the averaging filter.

2D methods

FourierFilterFlux.shearingLayer — Function

shearingLayer(inputSize::Union{Int,NTuple{N, T}}; scale = -1, shearLevel = scale, dType = Float32, σ = abs, trainable = false, plan = true, boundary=Pad(-1,-1), averagingLayer = false) where {N,T}

create a ConvFFT layer that uses shearlets from Shearlab.jl. By default it isn't trainable.

New Arguments

scale::Integer=-1: gives the number of scales to compute. If this is -1, it is computed automatically based on the image size, with 2 scales if the image is smaller than 24, and 4 otherwise -

shearLevel::Integer=scale: controls the number of shearing levels to use at each scale. Should roughly correspond to how far to shear in one direction at the finest scale.

averagingLayer:Bool=false: sets whether or not to only apply just the averaging filter.